requirements.txt
typer==0.12.3
httpx==0.27.0
python-dotenv==1.0.1
rich==13.7.1
